Fix bug with encodeAudioFile not generating out.ogg
Running the embed binary before wouldn't generate the encoded out.ogg
file for some reason. Ffmpeg was giving this error: "Format ogg
detected only with low score of 1, misdetection possible". So I checked
if the parameters were properly passed to the binary.

The problem was that c++ was blocking access to the audioFile (it was being opened,
closed and then passed as a parameter). This has been fixed with the
encodeAudioFile function.