Fix bug in encodeImage in EmbedAudio
EmbedAudio wasn't checking the filepath of the encodedAudioFile, but the
actual content of the encodedAudioFile itself. Then when the call to
encodedAudioFilePath was replaced with the path to the tempAudioFile in
the audioData struct, it was checking if "out.ogg" existed/isCorrupted
(but it was already renamed to "temp.ogg"). This has been patched up
by passing in a magic string.