Add autocomplete feature to oggex-qt
You can now autocomplete fields by passing them as command line
arguments. The only catch is that now the default message for passing
incorrect number of arguments is not available since calling argp_usage
outside of the parsing function isn't recommended. Additionally,
parse_opt doesn't fail anymore on passing incorrect arguments, so this
must be dealt with manually instead.

The main issue is that argp expects there to be an argument in argv.